Install both 32/64 bit shell extension
It's now possible to install both 32-bit and 64-bit version of the
shell extension on 64-bit systems.  32-bit version will be
registered in the registry view for 32-bit applications, and 64-bit
version will be registered in that for 64-bit applications.  Of
course only 32-bit version will be installed on 32-bit systems.